+++ app/templates/index.html
